Changed the OMSSA abundance scale from a long to a double (as detailed in OMSSA.mod.xsd). Moved project to version 1.4.7.

+++ src/main/java/de/proteinms/omxparser/tools/OmssaViewer.java
@@ -532,7 +532,7 @@ public class OmssaViewer extends javax.swing.JFrame {
 
                     chargeString = chargeString.replaceFirst("\\+", "");
 
-                    long omssaAbundanceScale = tempSpectrum.MSSpectrum_iscale;
+                    double omssaAbundanceScale = tempSpectrum.MSSpectrum_iscale;
 
                     ArrayList<Double> currentRealMzValues = new ArrayList();
                     ArrayList<Double> currentRealAbundanceValues = new ArrayList();
